+++ src/docs/HISTORY.md
@@ -44,6 +44,8 @@ v1.4.0
 - Fixed framing when loading a session file.
 - Improved startup times from the command-line and when loading the first clip.
 - Added reporting of Studio location if MRV2_STUDIOPATH or STUDIOPATH are set.
+- Fixed saving of movies in Vulkan.
+- Fixed saving of movies at half and quarter resolution in Vulkan.
 
 
 v1.3.9